mlops-zoomcamp-experiment-tracking/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-zoomcamp-experiment-tracking"
version = "0.1.0"
description = "MLOps Zoomcamp Experiment Tracking Application"
authors = [
    { name = "Nuno Paiva", email = "nuno@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "PyYAML==6.0.2",
    "optuna==4.3.0",
    "mlflow==2.17.2",
    "scikit-learn==1.3.2",
    "click==8.1.8",
    "pandas==2.0.3",
    "numpy==1.24.4",
    "interpret==0.6.6"
]

[project.optional-dependencies]
dev = [
    "seaborn==0.13.2",
    "nbformat==5.10.4",
    "plotly==5.22.0",
    "ipython==8.21.0",
    "notebook==6.5.7"
]
